91 Update all notebooks to MONAI v0.4 (#92)
* [DLMED] update to monai v0.4


* [DLMED] fix minor issues


* [DLMED] fix minor issues in examples


* [DLMED] fix typo


* [DLMED] fix MedNIST GAN issue

+++ 2d_segmentation/torch/unet_evaluation_array.py
@@ -74,7 +74,7 @@ def main(tempdir):
             value, _ = dice_metric(y_pred=val_outputs, y=val_labels)
             metric_count += len(value)
             metric_sum += value.item() * len(value)
-            saver.save_batch(val_outputs)
+            saver.save_batch(val_outputs.to(dtype=torch.int))
         metric = metric_sum / metric_count
         print("evaluation metric:", metric)
